+++ .gitignore
@@ -1,3 +1,4 @@
+package-lock.json
 node_modules
 demos/node_modules
 demos/package-lock.json
