+++ demos/package.json
@@ -14,6 +14,6 @@
   },
   "scripts": {
     "simple": "node ./simple/index.js",
-    "web": "node ./webclient/index.js"
+    "web": "node ./web/index.js"
   }
 }
