+++ demos/web/ClientClass.js
@@ -1,0 +1,18 @@
+class ClientClass {
+  /**
+   * ...
+   */
+  constructor() {
+    console.log("client> created");
+  }
+
+  /**
+   * ...
+   */
+  onConnect() {
+    console.log("client> connected to server");
+  }
+
+}
+
+module.exports = ClientClass;
