+++ demos/web/public/index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>Web demo</title>
+        <script src="socketless.js" async defer></script>
+        <script src="page.js" type="module" async defer></script>
+    </head>
+    <body>
+        <!-- ... -->
+    </body>
+</html>
