+++ demos/web/public/page.js
@@ -1,0 +1,2 @@
+import { WebClient } from "./ui.js";
+ClientServer.generateClientServer(WebClient);
