+++ demos/web/public/ui.js
@@ -1,0 +1,11 @@
+class WebClient {
+  constructor() {
+    // ...
+  }
+
+  update(state) {
+    // ...
+  }
+}
+
+export { WebClient };
