--- demos/webclient/public/index.html
@@ -1,2 +1,0 @@
-<script src="socketless.js"></script>
-<script src="start.js" type="module" async defer></script>
